Moved hardware specific parts into 'transponder_ir.c'.
