+++ src/main/drivers/transponder_ir.c
@@ -24,7 +24,12 @@
 #include "dma.h"
 #include "nvic.h"
 #include "io.h"
+#include "rcc.h"
 #include "timer.h"
+#if defined(STM32F4)
+#include "timer_stm32f4xx.h"
+#endif
+
 #include "transponder_ir.h"
 
 /*
@@ -38,6 +43,135 @@
 uint8_t transponderIrDMABuffer[TRANSPONDER_DMA_BUFFER_SIZE];
 
 volatile uint8_t transponderIrDataTransferInProgress = 0;
+
+static IO_t transponderIO = IO_NONE;
+static TIM_TypeDef *timer = NULL;
+#if defined(STM32F3)
+static DMA_Channel_TypeDef *dmaChannel = NULL;
+#elif defined(STM32F4)
+static DMA_Stream_TypeDef *stream = NULL;
+#else
+#error "Transponder not supported on this MCU."
+#endif
+
+static void TRANSPONDER_DMA_IRQHandler(dmaChannelDescriptor_t* descriptor)
+{
+    if (DMA_GET_FLAG_STATUS(descriptor, DMA_IT_TCIF)) {
+        transponderIrDataTransferInProgress = 0;
+#if defined(STM32F3)
+        DMA_Cmd(descriptor->channel, DISABLE);
+#elif defined(STM32F4)
+        DMA_Cmd(descriptor->stream, DISABLE);
+#endif
+        DMA_CLEAR_FLAG(descriptor, DMA_IT_TCIF);
+    }
+}
+
+void transponderIrHardwareInit(ioTag_t ioTag)
+{
+    if (!ioTag) {
+        return;
+    }
+
+    TIM_TimeBaseInitTypeDef  TIM_TimeBaseStructure;
+    TIM_OCInitTypeDef  TIM_OCInitStructure;
+    DMA_InitTypeDef DMA_InitStructure;
+
+    const timerHardware_t *timerHardware = timerGetByTag(ioTag, TIM_USE_ANY);
+    timer = timerHardware->tim;
+
+#if defined(STM32F3)
+    if (timerHardware->dmaChannel == NULL) {
+        return;
+    }
+#elif defined(STM32F4)
+    if (timerHardware->dmaStream == NULL) {
+        return;
+    }
+#endif
+
+    transponderIO = IOGetByTag(ioTag);
+    IOInit(transponderIO, OWNER_TRANSPONDER, 0);
+    IOConfigGPIOAF(transponderIO, IO_CONFIG(GPIO_Mode_AF, GPIO_Speed_50MHz, GPIO_OType_PP, GPIO_PuPd_DOWN), timerHardware->alternateFunction);
+
+    dmaInit(timerHardware->dmaIrqHandler, OWNER_TRANSPONDER, 0);
+    dmaSetHandler(timerHardware->dmaIrqHandler, TRANSPONDER_DMA_IRQHandler, NVIC_PRIO_TRANSPONDER_DMA, 0);
+
+    RCC_ClockCmd(timerRCC(timer), ENABLE);
+
+    /* Time base configuration */
+    TIM_TimeBaseStructInit(&TIM_TimeBaseStructure);
+    TIM_TimeBaseStructure.TIM_Period = 156;
+    TIM_TimeBaseStructure.TIM_Prescaler = 0;
+    TIM_TimeBaseStructure.TIM_ClockDivision = 0;
+    TIM_TimeBaseStructure.TIM_CounterMode = TIM_CounterMode_Up;
+    TIM_TimeBaseInit(timer, &TIM_TimeBaseStructure);
+
+    /* PWM1 Mode configuration: Channel1 */
+    TIM_OCStructInit(&TIM_OCInitStructure);
+    TIM_OCInitStructure.TIM_OCMode = TIM_OCMode_PWM1;
+    if (timerHardware->output & TIMER_OUTPUT_N_CHANNEL) {
+        TIM_OCInitStructure.TIM_OutputNState = TIM_OutputNState_Enable;
+        TIM_OCInitStructure.TIM_OCNIdleState = TIM_OCNIdleState_Reset;
+    } else {
+        TIM_OCInitStructure.TIM_OutputState = TIM_OutputState_Enable;
+        TIM_OCInitStructure.TIM_OCIdleState = TIM_OCIdleState_Set;
+    }
+
+    TIM_OCInitStructure.TIM_OCPolarity =  (timerHardware->output & TIMER_OUTPUT_INVERTED) ? TIM_OCPolarity_Low : TIM_OCPolarity_High;
+    TIM_OCInitStructure.TIM_Pulse = 0;
+#if defined(STM32F3)
+    TIM_OC1Init(timer, &TIM_OCInitStructure);
+    TIM_OC1PreloadConfig(timer, TIM_OCPreload_Enable);
+#elif defined(STM32F4)
+    timerOCInit(timer, timerHardware->channel, &TIM_OCInitStructure);
+    timerOCPreloadConfig(timer, timerHardware->channel, TIM_OCPreload_Enable);
+#endif
+    TIM_CtrlPWMOutputs(timer, ENABLE);
+
+    /* configure DMA */
+#if defined(STM32F3)
+    dmaChannel = timerHardware->dmaChannel;
+    DMA_DeInit(dmaChannel);
+#elif defined(STM32F4)
+    stream = timerHardware->dmaStream;
+    DMA_Cmd(stream, DISABLE);
+    DMA_DeInit(stream);
+#endif
+
+    DMA_StructInit(&DMA_InitStructure);
+    DMA_InitStructure.DMA_PeripheralBaseAddr = (uint32_t)timerCCR(timer, timerHardware->channel);
+#if defined(STM32F3)
+    DMA_InitStructure.DMA_MemoryBaseAddr = (uint32_t)transponderIrDMABuffer;
+#elif defined(STM32F4)
+    DMA_InitStructure.DMA_Memory0BaseAddr = (uint32_t)transponderIrDMABuffer;
+#endif
+    DMA_InitStructure.DMA_BufferSize = TRANSPONDER_DMA_BUFFER_SIZE;
+    DMA_InitStructure.DMA_PeripheralInc = DMA_PeripheralInc_Disable;
+    DMA_InitStructure.DMA_MemoryInc = DMA_MemoryInc_Enable;
+    DMA_InitStructure.DMA_PeripheralDataSize = DMA_PeripheralDataSize_HalfWord;
+    DMA_InitStructure.DMA_MemoryDataSize = DMA_MemoryDataSize_Byte;
+    DMA_InitStructure.DMA_Mode = DMA_Mode_Normal;
+    DMA_InitStructure.DMA_Priority = DMA_Priority_High;
+#if defined(STM32F3)
+    DMA_InitStructure.DMA_DIR = DMA_DIR_PeripheralDST;
+    DMA_InitStructure.DMA_M2M = DMA_M2M_Disable;
+
+    DMA_Init(dmaChannel, &DMA_InitStructure);
+#elif defined(STM32F4)
+    DMA_InitStructure.DMA_DIR = DMA_DIR_MemoryToPeripheral;
+
+    DMA_Init(stream, &DMA_InitStructure);
+#endif
+
+    TIM_DMACmd(timer, timerDmaSource(timerHardware->channel), ENABLE);
+
+#if defined(STM32F3)
+    DMA_ITConfig(dmaChannel, DMA_IT_TC, ENABLE);
+#elif defined(STM32F4)
+    DMA_ITConfig(stream, DMA_IT_TC, ENABLE);
+#endif
+}
 
 bool transponderIrInit(void)
 {
@@ -57,6 +191,7 @@ bool transponderIrInit(void)
 
 
     transponderIrHardwareInit(ioTag);
+
     return true;
 }
 
@@ -119,6 +254,40 @@ void transponderIrUpdateData(const uint8_t* transponderData)
     updateTransponderDMABuffer(transponderData);
 }
 
+void transponderIrDMAEnable(void)
+{
+#if defined(STM32F3)
+    DMA_SetCurrDataCounter(dmaChannel, TRANSPONDER_DMA_BUFFER_SIZE);  // load number of bytes to be transferred
+#elif defined(STM32F4)
+    DMA_SetCurrDataCounter(stream, TRANSPONDER_DMA_BUFFER_SIZE);  // load number of bytes to be transferred
+#endif
+    TIM_SetCounter(timer, 0);
+    TIM_Cmd(timer, ENABLE);
+#if defined(STM32F3)
+    DMA_Cmd(dmaChannel, ENABLE);
+#elif defined(STM32F4)
+    DMA_Cmd(stream, ENABLE);
+#endif
+}
+
+void transponderIrDisable(void)
+{
+#if defined(STM32F3)
+    DMA_Cmd(dmaChannel, DISABLE);
+#elif defined(STM32F4)
+    DMA_Cmd(stream, DISABLE);
+#endif
+    TIM_Cmd(timer, DISABLE);
+
+    IOInit(transponderIO, OWNER_TRANSPONDER, 0);
+    IOConfigGPIOAF(transponderIO, IO_CONFIG(GPIO_Mode_AF, GPIO_Speed_50MHz, GPIO_OType_PP, GPIO_PuPd_DOWN), timerHardware->alternateFunction);
+
+#ifdef TRANSPONDER_INVERTED
+    IOHi(transponderIO);
+#else
+    IOLo(transponderIO);
+#endif
+}
 
 void transponderIrTransmit(void)
 {
